requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
colorama==0.4.1
Django==2.2.7
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-tinymce==2.8.0
importlib-metadata==0.23
more-itertools==7.2.0
mysqlclient==1.4.4
packaging==19.2
Pillow==6.2.1
pip==20.0.2
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.3
pytest==5.2.2
pytz==2019.3
setuptools==41.2.0
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
zipp==0.6.0
